+++ static/js/pages/instances/forms/CpuLimitSelector.tsx
