+++ static/js/pages/instances/forms/MemoryLimitSelector.tsx
